requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.8
idna==3.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
